fix(contracts): reject a fractional users[].startOffsetMs
LiveDeltaSchema.users[].startOffsetMs was a bare z.number() while
UsersResponseSchema's identically-named field has .int() -- the same
count-field inconsistency Task 2 fixed for everything else on this
schema, missed only because it fell outside that finding's citation.
buildDelta always emits it from a UserBucket's own integer offset, so
this only tightens validation to match what the field actually holds.

+++ packages/contracts/src/live-delta.ts
@@ -96,7 +96,7 @@ export const LiveDeltaSchema = z.object({
   users: z.array(
     z.object({
       scenario: z.string(),
-      startOffsetMs: z.number(),
+      startOffsetMs: z.number().int(),
       active: z.number(),
     }),
   ),
